pyproject.toml
[project]
name = "qugenai"
version = "0.1.0"
description = "AI-powered exam content generator"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "nanodjango>=0.10.0",
    "openai>=1.66.5",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "ruff>=0.11.0",
]

[tool.ruff]
line-length = 88
indent-width = 4

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
